Add files via upload
This solution includes two small demo apps:
"Eine kleine Lern-App" is a small PoC to check whether Power Apps are suitable for writing learning applications for primary school children.
The “Zeiträuber” is a small app to log your daily routine. It serves as a demonstrator of how easy it is to create small apps.
Both apps require a small SharePoint infrastructure, which is created via the "Create SP-Infrastructure" solution.